requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
apscheduler>=3.10.0
openpyxl>=3.1.0
pyodbc>=5.0.0
cryptography>=41.0.0
jinja2>=3.1.0
pyyaml>=6.0.0
